fix(serve): keep write provenance off startup bundle
