+++ packages/cli/vitest.config.ts
@@ -27,6 +27,10 @@ export default defineConfig({
         __dirname,
         '../core/src/memory/scopes.ts',
       ),
+      '@qwen-code/qwen-code-core/toolWriteOrigin': path.resolve(
+        __dirname,
+        '../core/src/services/tool-write-origin.ts',
+      ),
       '@qwen-code/qwen-code-core': path.resolve(__dirname, '../core/index.ts'),
       // cli's daemon-status-provider.test.ts imports `FakeAgent` /
       // `makeChannel` from acp-bridge's package-private
