+++ packages/core/src/services/fileSystemService.test.ts
@@ -12,8 +12,6 @@ import {
   resetUtf8BomCache,
   detectLineEnding,
   ensureCrlfLineEndings,
-  buildToolWriteOriginMeta,
-  parseToolWriteOriginMeta,
 } from './fileSystemService.js';
 import { encodeTextFileContent } from '../utils/sync-file-encoding.js';
 
@@ -59,54 +57,6 @@ vi.mock('../utils/fileUtils.js', async (importOriginal) => {
 });
 
 import { readFileWithLineAndLimit } from '../utils/fileUtils.js';
-
-describe('tool write origin metadata', () => {
-  it('round-trips each supported source', () => {
-    for (const source of [
-      'write_file',
-      'edit',
-      'notebook_edit',
-      'shell_sed_edit',
-    ] as const) {
-      const meta = buildToolWriteOriginMeta({ bom: true }, source);
-      expect(parseToolWriteOriginMeta(meta)).toBe(source);
-      expect(meta?.['bom']).toBe(true);
-    }
-  });
-
-  it('replaces caller-supplied provenance', () => {
-    const meta = buildToolWriteOriginMeta(
-      {
-        'qwen-code/tool-write-origin': {
-          version: 1,
-          source: 'write_file',
-        },
-      },
-      'edit',
-    );
-
-    expect(meta?.['qwen-code/tool-write-origin']).toEqual({
-      version: 1,
-      source: 'edit',
-    });
-  });
-
-  it.each([
-    undefined,
-    null,
-    [],
-    { version: 2, source: 'write_file' },
-    { version: 1, source: 'unknown' },
-    { version: 1, source: 'write_file', extra: true },
-    { source: 'write_file' },
-  ])('rejects malformed marker %j', (marker) => {
-    expect(
-      parseToolWriteOriginMeta({
-        'qwen-code/tool-write-origin': marker,
-      }),
-    ).toBeUndefined();
-  });
-});
 
 describe('StandardFileSystemService', () => {
   let fileSystem: StandardFileSystemService;
